+++ src/packages/piral-core/src/state/createActions.ts
@@ -1,15 +1,6 @@
 import * as actions from '../actions';
 import { Atom } from '@dbeining/react-atom';
-import { PiralLoadingIndicator } from '../components';
-import { renderInDom, convertComponent } from '../utils';
-import {
-  GlobalState,
-  GlobalStateContext,
-  EventEmitter,
-  PiralDefineActions,
-  ForeignComponent,
-  ComponentContext,
-} from '../types';
+import { GlobalState, GlobalStateContext, EventEmitter, PiralDefineActions } from '../types';
 
 function createContext(state: Atom<GlobalState>, events: EventEmitter) {
   const ctx = {
@@ -17,32 +8,6 @@ function createContext(state: Atom<GlobalState>, events: EventEmitter) {
     apis: {},
     converters: {
       html: ({ component }) => component,
-      lazy: ({ load }) => {
-        let present: [HTMLElement, any, ComponentContext] = undefined;
-        let portalId: string = undefined;
-        const promise = load.current || (load.current = load().then(c => convertComponent(ctx.converters[c.type], c)));
-        const component: ForeignComponent<any> = {
-          mount(...args) {
-            portalId = renderInDom(ctx, args[0], PiralLoadingIndicator, {});
-            present = args;
-          },
-          update(...args) {
-            present = args;
-          },
-          unmount() {
-            portalId = undefined;
-            present = undefined;
-          },
-        };
-        promise.then(({ mount, unmount, update }) => {
-          portalId && ctx.destroyPortal(portalId);
-          component.mount = mount;
-          component.unmount = unmount;
-          component.update = update;
-          present && mount(...present);
-        });
-        return component;
-      },
     },
     state,
   } as GlobalStateContext;